// src/app/product/models/product.model.ts
export interface ProductModel {
  BrandId: number;
  BrandName: string;
  CreatedOnUtc: string;
  HTSDescription: string;
  HTSNumber: number;
  ProductGuid: number;
  ProductName: string;
  ShortDescription: string;
  Status: string;
  StatusId: number;
  FAQ?: any;
  PicPath: string;
  category: string;
  StatusComment: any;
  Tag?: any;
  Color?: string;
}