// src/app/product/models/category-attribute.model.ts
export interface CategoryAttributeModel {
  AttributeGroupId: number;
  AttributeId: number;
  AttributeName: string;
  CategoryId: number;
  CategoryName: string;
  DefaultValues: string[];
  DefaultValueIds: number[];
  GroupName: string;
  Id: number;
  InsertValueManually: boolean;
  IsRequired: boolean;
  UnitId: number;
  UnitName: string;
}